requirements.txt
streamlit>=1.41.0
pandas>=2.0.0
yfinance>=0.2.28
plotly>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0